Refocus to be independent crate
