+++ nym-vpn-core/crates/nym-vpn-store/src/persistence/mod.rs
@@ -1,0 +1,5 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+pub mod ephemeral;
+pub mod on_disk;
